requirements.txt
flask
pyswisseph
gunicorn
python-dotenv
svgwrite
geopy
timezonefinder
pytz
